+++ src/interface.h
@@ -11,7 +11,7 @@
 #define NEED_SERIAL
 
 void lcd_display(uint8_t x, uint8_t y, char buffer[33]);
-void serial_display(double rpm);
-void serial_rpm(double rpm, double hertz);
+void serial_display(float rpm);
+void serial_rpm(float rpm, float hertz);
 
 #endif
